+++ spec/javascripts/govuk_publishing_components/lib/cookie-settings-spec.js
@@ -17,8 +17,8 @@ describe('cookieSettings', function () {
       '<form data-module="cookie-settings">' +
         '<input type="radio" id="settings-on" name="cookies-settings" value="on">' +
         '<input type="radio" id="settings-off" name="cookies-settings" value="off">' +
-        '<input type="radio" name="cookies-usage" value="on">' +
-        '<input type="radio" name="cookies-usage" value="off">' +
+        '<input type="radio" id="usage-on" name="cookies-usage" value="on">' +
+        '<input type="radio" id="usage-off" name="cookies-usage" value="off">' +
         '<input type="radio" name="cookies-campaigns" value="on">' +
         '<input type="radio" name="cookies-campaigns" value="off">' +
         '<button id="submit-button" type="submit">Submit</button>' +
@@ -136,10 +136,13 @@ describe('cookieSettings', function () {
       element.querySelector('#settings-on').checked = false
       element.querySelector('#settings-off').checked = true
 
+      element.querySelector('#usage-on').checked = true
+      element.querySelector('#usage-off').checked = false
+
       var button = element.querySelector('#submit-button')
       button.click()
 
-      expect(GOVUK.analytics.trackEvent).toHaveBeenCalledWith('cookieSettings', 'Save changes', { label: 'settings-no usage-no campaigns-no ' })
+      expect(GOVUK.analytics.trackEvent).toHaveBeenCalledWith('cookieSettings', 'Save changes', { label: 'settings-no usage-yes campaigns-no ' })
     })
   })
 
@@ -193,6 +196,45 @@ describe('cookieSettings', function () {
       button.click()
 
       expect(confirmationMessage.style.display).toEqual('block')
+    })
+  })
+
+  describe('when setting usage cookies to false', function () {
+    beforeEach(function () {
+      spyOn(GOVUK.analyticsGa4.core, 'getGemVersion').and.returnValue('aVersion')
+    })
+    it('stops analytics tracking', function () {
+      var analytics = new GOVUK.Analytics({
+        universalId: 'universal-id',
+        cookieDomain: '.www.gov.uk',
+        siteSpeedSampleRate: 100
+      })
+      window.dataLayer = []
+
+      // Expect the default return value when you send an analytics event
+      expect(analytics.trackEvent()).toEqual(undefined)
+      expect(analytics.trackShare()).toEqual(undefined)
+      expect(analytics.trackPageview()).toEqual(undefined)
+      expect(GOVUK.analyticsGa4.core.sendData({})).toEqual(undefined)
+
+      new GOVUK.Modules.CookieSettings(element).init()
+
+      element.querySelector('#usage-on').checked = false
+      element.querySelector('#usage-off').checked = true
+
+      var button = element.querySelector('#submit-button')
+      button.click()
+
+      expect(window.GOVUK.stopSendingAnalytics).toEqual(true)
+
+      // Expect a false return value now when you send an analytics event
+
+      expect(analytics.trackEvent()).toEqual(false)
+      expect(analytics.trackShare()).toEqual(false)
+      expect(analytics.trackPageview()).toEqual(false)
+      expect(window.GOVUK.analyticsGa4.core.sendData()).toEqual(false)
+
+      expect(window.GOVUK.LUX).toEqual({})
     })
   })
 })